Restructure the files into folders
