+++ src/components/Expenses/Expenses.js
@@ -1,8 +1,8 @@
 import ExpenseItem from './ExpenseItem';
-import Card from './Card';
+import Card from '../UI/Card';
 import './Expenses.css';
 
-function Expenses(props) {
+const Expenses = (props) => {
   const items = props.items;
 
   return (
