+++ src/components/UI/Card.js
@@ -1,6 +1,6 @@
 import './Card.css';
 
-function Card(props) {
+const Card = (props) => {
   const classes = 'card ' + props.className;
 
   return (
